+++ commons-jms-mq/src/main/java/co/com/bancolombia/commons/jms/mq/utils/MQUtils.java
@@ -62,7 +62,7 @@ public class MQUtils {
     }
 
     public static void setQMNameIfNotSet(JMSContext context, Queue queue, boolean useTemporaryQueue) {
-        if (((MQQueue) queue).getBaseQueueManagerName() != null) {
+        if (((MQQueue) queue).getBaseQueueManagerName() == null) {
             String qmName = useTemporaryQueue ? extractQMNameWithTempQueue(context) : extractQMName(context);
             setQMName(queue, qmName);
         }
